+++ cmd/hidden_lake/helpers/traffic/internal/config/types.go
@@ -18,7 +18,7 @@ type IConfigSettings interface {
 	message.ISettings
 	net_message.ISettings
 
-	GetQueuePeriodMS() uint64
+	GetMaxQueuePeriodMS() uint64
 	GetMessagesCapacity() uint64
 	GetLimitVoidSizeBytes() uint64
 	GetStorageEnabled() bool