+++ cmd/hidden_lake/helpers/traffic/internal/handler/config_settings.go
@@ -26,7 +26,7 @@ func HandleConfigSettingsAPI(pCfg config.IConfig, pLogger logger.ILogger) http.H
 		_ = api.Response(pW, http.StatusOK, config.SConfigSettings{
 			FMessageSizeBytes:   sett.GetMessageSizeBytes(),
 			FWorkSizeBits:       sett.GetWorkSizeBits(),
-			FQueuePeriodMS:      sett.GetQueuePeriodMS(),
+			FMaxQueuePeriodMS:   sett.GetMaxQueuePeriodMS(),
 			FKeySizeBits:        sett.GetKeySizeBits(),
 			FLimitVoidSizeBytes: sett.GetLimitVoidSizeBytes(),
 			FMessagesCapacity:   sett.GetMessagesCapacity(),