+++ cmd/hidden_lake/helpers/traffic/internal/handler/config_settings_test.go
@@ -50,7 +50,7 @@ func TestHandleConfigSettingsAPI(t *testing.T) {
 		return
 	}
 
-	if settings.GetQueuePeriodMS() != hls_settings.CDefaultQueuePeriod {
+	if settings.GetMaxQueuePeriodMS() != hls_settings.CDefaultQueuePeriod {
 		t.Error("incorrect queue period ms")
 		return
 	}