+++ cmd/hidden_lake/helpers/traffic/internal/handler/handler_test.go
@@ -119,7 +119,7 @@ func testRunService(db database.IDatabase, addr string, addrNode string) (*http.
 		FSettings: &config.SConfigSettings{
 			FMessageSizeBytes:   testutils.TCMessageSize,
 			FWorkSizeBits:       testutils.TCWorkSize,
-			FQueuePeriodMS:      hls_settings.CDefaultQueuePeriod,
+			FMaxQueuePeriodMS:   hls_settings.CDefaultQueuePeriod,
 			FLimitVoidSizeBytes: hls_settings.CDefaultLimitVoidSize,
 			FKeySizeBits:        testutils.TcKeySize,
 			FNetworkKey:         testutils.TCNetworkKey,